Fix all example tests
Some tests were not running, because of their dependencies (HomeController).

The tests have an optimal reading order, because each of them builds on concepts explained on the one before. I renamed them to reflect that.

I also wrote their explanations. The source of everything explained is this: https://spring.io/guides/gs/testing-web/